+++ pinus-cluster/src/main/java/org/pinus4j/cache/impl/AbstractMemCachedCache.java
@@ -22,7 +22,6 @@ import java.util.List;
 
 import net.spy.memcached.MemcachedClient;
 
-import org.pinus4j.cache.ICache;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -32,50 +31,39 @@ import org.slf4j.LoggerFactory;
  * @author duanbn
  * @since 0.7.1
  */
-public abstract class AbstractMemCachedCache implements ICache {
+public abstract class AbstractMemCachedCache extends AbstractCache {
 
     public static final Logger LOG = LoggerFactory.getLogger(AbstractMemCachedCache.class);
 
-	protected MemcachedClient memClient;
+    protected MemcachedClient  memClient;
 
-	protected int expire = 30;
+    public AbstractMemCachedCache(String address, int expire) {
+        super(address, expire);
 
-	public AbstractMemCachedCache(String address, int expire) {
-		this.expire = expire;
+        try {
+            List<InetSocketAddress> servers = new ArrayList<InetSocketAddress>();
 
-		try {
-			List<InetSocketAddress> servers = new ArrayList<InetSocketAddress>();
+            String[] addresses = address.split(",");
 
-			String[] addresses = address.split(",");
+            InetSocketAddress socketAddress = null;
+            for (String addr : addresses) {
+                String[] pair = addr.split(":");
+                socketAddress = new InetSocketAddress(pair[0], Integer.parseInt(pair[1]));
+                servers.add(socketAddress);
+            }
 
-			InetSocketAddress socketAddress = null;
-			for (String addr : addresses) {
-				String[] pair = addr.split(":");
-				socketAddress = new InetSocketAddress(pair[0], Integer.parseInt(pair[1]));
-				servers.add(socketAddress);
-			}
+            this.memClient = new MemcachedClient(servers);
+        } catch (Exception e) {
+            throw new RuntimeException("连接memcached服务器失败", e);
+        }
+    }
 
-			this.memClient = new MemcachedClient(servers);
-		} catch (Exception e) {
-			throw new RuntimeException("连接memcached服务器失败", e);
-		}
-
-		if (expire > 0) {
-			this.expire = expire;
-		}
-	}
-
-	@Override
-	public int getExpire() {
-		return this.expire;
-	}
-
-	/**
-	 * 销毁对象
-	 */
-	@Override
-	public void close() {
-		this.memClient.shutdown();
-	}
+    /**
+     * 销毁对象
+     */
+    @Override
+    public void close() {
+        this.memClient.shutdown();
+    }
 
 }
